+++ Scripts/setup_dependencies.sh
@@ -38,7 +38,8 @@ if [[ $dependencies == *"pyenv"* ]]; then
     export PATH="$PYENV_ROOT/bin":$PATH;
     eval "$(pyenv init -)"
     # get latest python version from the pyenv list
-    LATEST_PYTHON=$(pyenv install --list | grep --extended-regexp "^\s*[0-9][0-9.]*[0-9]\s*$" | tail -1);
-    pyenv install "$LATEST_PYTHON" 2.7.18;
-    pyenv global "$LATEST_PYTHON" 2.7.18;
+    LATEST_PYTHON=$(pyenv install --list | grep --extended-regexp "^\s*[0-9][0-9.]*[0-9]\s*$" | tail -1 | xargs);
+    pyenv install $LATEST_PYTHON --force
+    pyenv install 2.7.18 --force
+    pyenv global $LATEST_PYTHON 2.7.18;
 fi